apps/backend/requirements.txt
# Auto-Build Framework Dependencies
# SDK 0.1.25+ required for improved tool use concurrency handling
# Earlier versions had 400 errors when tool_use blocks had partial failures
claude-agent-sdk>=0.1.25
python-dotenv>=1.0.0

# TOML parsing fallback for Python < 3.11
tomli>=2.0.0; python_version < "3.11"

# Linux Secret Service support for credential storage
# Provides access to the Freedesktop.org Secret Service API via DBus
# Used on Linux to store OAuth tokens in gnome-keyring/kwallet
secretstorage>=3.3.3; sys_platform == "linux"

# Memory Integration - LadybugDB (embedded graph database)
# Requires Python 3.12+ (no Docker required)
real_ladybug>=0.13.0; python_version >= "3.12"
graphiti-core>=0.5.0; python_version >= "3.12"
# pandas is required by real_ladybug for get_as_df() method
# pandas 2.2.0+ required for pre-built wheels on Python 3.12
pandas>=2.2.0; python_version >= "3.12"

# Windows-specific dependency for LadybugDB/Graphiti
# pywin32 provides Windows system bindings required by real_ladybug
# Required on all Python versions on Windows (ACS-306) - MCP library unconditionally imports win32api
pywin32>=306; sys_platform == "win32"

# Google AI (optional - for Gemini LLM and embeddings)
google-generativeai>=0.8.0
openai>=1.0.0

# Pydantic for structured output schemas
pydantic>=2.0.0

# Error tracking (optional - requires SENTRY_DSN environment variable)
sentry-sdk>=2.0.0
